Termination Tolerance Option
Most of the times the solution to a MIP problem does not have to be the
absolute best possible solution, but rather a solution relatively close
to the relaxed one. High-level modelling systems like GAMS therefore
introduce a so called termination tolerance, where the solver will stop
the solution process when the proportional difference between the
solution found and the best theoretical objective function is
guaranteed to be smaller than the given termination tolerance.
I tried to add this option to the jsLPSolver, which can be used by
passing the option „optcr“ as follows:
```javascript
results = solver.Solve(model, 0.5);
```
This generally reduces the solve time for large scale MIPs
significantly.
The default value of Optcr is set to 0.1.